% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADF.R
\name{ADF}
\alias{ADF}
\title{Return the ADF statistic}
\usage{
ADF(y, IC, adflag)
}
\arguments{
\item{y}{A vector, The data.}

\item{IC}{A positive integer. 0 for fixed lag order 1 for AIC and 2 for BIC.}

\item{adflag}{A positive integer. Lag order when IC=0; maximum number of lags when
IC>0.}
}
\value{
Numeric, ADF test statistic.
}
\description{
\code{ADF} calculates the augmented Dickey-Fuller (ADF) test
  statistic with lag order set fixed or selected by AIC or BIC.
}
\examples{
\donttest{
y <- rnorm(100)
ADFstat <- ADF(y,  IC = 0, adflag = 1)
}
}
\references{
Said, S. E., & Dickey, D. A. (1984). Testing for Unit Roots in
  ARMA Models of Unknown Order. \emph{Biometrika}, 71(1984), 599--607.
}
