% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disp.R
\name{disp}
\alias{disp}
\title{Display bubble/crisis periods}
\usage{
disp(OT)
}
\arguments{
\item{OT}{A vector. Dates identified as bubbles/crisis by the
\code{spymonitor::locate} function.}
}
\value{
A vector of strings with bubble/crisis periods.
}
\description{
TBA
}
\examples{
\donttest{
y  <- rnorm(100)
}
}
\references{
Phillips, P. C. B., Shi, S., & Yu, J. (2015a). Testing for
  multiple bubbles: Historical episodes of exuberance and collapse in the S&P
  500. \emph{International Economic Review}, 56(4), 1034--1078. Phillips, P.
  C. B., Shi, S., & Yu, J. (2015b). Testing for multiple bubbles: Limit
  Theory for Real-Time Detectors. \emph{International Economic Review},
  56(4), 1079--1134.
}
