% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate}
\alias{locate}
\title{Locate periods with non-zero bubble indicator}
\usage{
locate(ind, date)
}
\arguments{
\item{ind}{A vector. A dummy variable that equals 1 for a bubble/crisis
period and 0 otherwise.}

\item{index}{A vector. Dates of the time series.}
}
\value{
A vector. Dates identified as bubbles or crisis.
}
\description{

}
\examples{
\donttest{
y <- rnorm(100)
}
}
\references{
Phillips, P. C. B., Shi, S., & Yu, J. (2015a). Testing for
  multiple bubbles: Historical episodes of exuberance and collapse in the S&P
  500. \emph{International Economic Review}, 56(4), 1034--1078. Phillips, P.
  C. B., Shi, S., & Yu, J. (2015b). Testing for multiple bubbles: Limit
  Theory for Real-Time Detectors. \emph{International Economic Review},
  56(4), 1079--1134.
}
