% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psymonitor-package.R
\docType{package}
\name{psymonitor-package}
\alias{psymonitor}
\alias{psymonitor-package}
\title{psymonitor: Real Time Monitoring of Bubbles and Crisis with R}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

TBA
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itamarcaspi/psymonitor}
  \item Report bugs at \url{https://github.com/itamarcaspi/psymonitor/issues}
}

}
\author{
\strong{Maintainer}: Itamar Caspi \email{caspi.itamar@gmail.com}

Authors:
\itemize{
  \item Peter C.B. Phillips
  \item Shuping Shi
}

}
\keyword{internal}
