% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spread}
\alias{spread}
\title{Credit risk in the European sovereign sector}
\format{A tibble}
\source{
Datastream International.
}
\usage{
spread
}
\description{
Credit risk in the European sovereign sector: June 1997--June 2016. Proxied by
an index constructed as a GDP weighted 10-year government bond yield of the
GIIPS (Greece, Ireland, Italy, Portugal, and Spain) countries.
}
\examples{
head(spread)

}
\keyword{datasets}
