% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wmboot.R
\name{wmboot}
\alias{wmboot}
\title{Conduct the new composite bootstrapping for the PSY test.}
\usage{
wmboot(y, swindow0, IC, adflag, Tb, nboot)
}
\arguments{
\item{y}{A vector. The data.}

\item{swindow0}{minimum window size.}

\item{IC}{A positive integer. 0 for fixed lag order 1 for AIC and 2 for BIC.}

\item{adflag}{A positive integer. Lag order when IC=0; maximum number of lags when IC>0.}

\item{Tb}{A positive integer. The simulated sample size swindow0+ controlling.}

\item{nboot}{A positive integer. Number of bootstrap replications.}
}
\value{
TBA
}
\description{
\code{wmboot} TBA
}
\examples{
\donttest{
y <- rnorm(100)
wmboot(y,  swindow0 = 10, IC = 0, adflag = 1, Tb = 10, nboot = 99)
}
}
\references{
Phillips, P. C. B., Shi, S., & Yu, J. (2015a). Testing for multiple bubbles:
  Historical episodes of exuberance and collapse in the S&P 500.
  \emph{International Economic Review}, 56(4), 1034--1078.
  Phillips, P. C. B., Shi, S., & Yu, J. (2015b). Testing for multiple bubbles:
  Limit Theory for Real-Time Detectors.
  \emph{International Economic Review}, 56(4), 1079--1134.
}
