% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psymonitor-package.R
\docType{package}
\name{psymonitor-package}
\alias{psymonitor}
\alias{psymonitor-package}
\title{psymonitor: psymonitor: Real Time Monitoring of Bubbles and Crisis
    with R}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The *psymonitor* package provides an R
    implementation of the popular real-time monitoring strategy proposed
    by Phillips, Shi and Yu (2015a,b;PSY), along with a new bootstrap
    procedure designed to mitigate the potential impact of
    heteroskedasticity and to effect family-wise size control in recursive
    testing algorithms (Phillips and Shi, forthcoming). This methodology
    has been shown effective for bubble and crisis detection (PSY,
    2015a,b; Phillips and Shi, 2017) and is now widely used by academic
    researchers, central bank economists, and fiscal regulators.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itamarcaspi/psymonitor}
  \item Report bugs at \url{https://github.com/itamarcaspi/psymonitor/issues}
}

}
\author{
\strong{Maintainer}: Itamar Caspi \email{caspi.itamar@gmail.com}

Authors:
\itemize{
  \item Peter C.B. Phillips
  \item Shuping Shi
}

}
\keyword{internal}
