% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psymonitor-package.R
\docType{package}
\name{psymonitor-package}
\alias{psymonitor}
\alias{psymonitor-package}
\title{psymonitor: Real Time Monitoring of Asset Markets: Bubbles and Crisis}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Apply the popular real-time monitoring strategy
    proposed by Phillips, Shi and Yu (2015a,b;PSY) <doi:10.1111/iere.12132>, <doi:10.1111/iere.12131>, along with a new
    bootstrap procedure designed to mitigate the potential impact of
    heteroskedasticity and to effect family-wise size control in recursive
    testing algorithms (Phillips and Shi, forthcoming).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/itamarcaspi/psymonitor}
  \item Report bugs at \url{https://github.com/itamarcaspi/psymonitor/issues}
}

}
\author{
\strong{Maintainer}: Itamar Caspi \email{caspi.itamar@gmail.com}

Authors:
\itemize{
  \item Peter C.B. Phillips
  \item Shuping Shi
}

}
\keyword{internal}
