% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp}
\alias{snp}
\title{S&P 500 price to dividend ratio}
\format{Data frame}
\source{
Datastream International.
}
\usage{
snp
}
\description{
S&P 500 price to dividend ratio: January 1973-- July 2017
}
\examples{
head(snp)

}
\keyword{datasets}
